Device53 VS Избавляемся от предупреждений

+++ sources/S8-53/src/Display/Display.h
@@ -164,7 +164,7 @@ private:
     
     static bool NeedForClearScreen();
     
-    static void WriteStringAndNumber(char *text, int16 x, int16 y, int number);
+    static void WriteStringAndNumber(char *text, int x, int y, int number);
     
     static void DrawGridType3(int left, int top, int right, int bottom, int centerX, int centerY, int deltaX, int deltaY, int stepX, int stepY);
