+++ sources/S8-53/src/VCP/SCPI/displaySCPI.cpp
@@ -308,7 +308,7 @@ void SCPI::DISPLAY::GRID_BRIGHTNESS(uint8 *buffer)
     int intVal = 0;
     if (SCPI::FirstIsInt(buffer, &intVal, 0, 100))
     {
-        BRIGHTNESS_GRID = intVal;
+        BRIGHTNESS_GRID = (int16)intVal;
         Display::RunAfterDraw(OnChanged_Grid_Brightness);
     }
     else
